refactored declarations
