+++ source/models/unity-context.ts
@@ -1,4 +1,4 @@
-import "../declarations/global";
+import "../typings/global";
 import IUnityConfig from "../interfaces/unity-config";
 import IUnityEvent from "../interfaces/unity-event";
 import IUnityContextEventMap from "../interfaces/unity-context-event-map";
